Final commit before submission
Improved beat_saber (single) by preventing the fictitious roller to spin around. The same improvement does not work well on the duo setup, so it is not used there. Also added the original version as beat_saber_og.py. This has the point-to-point goal without fictitious joints. Also small updates to urdf files.